+++ Classes/Domain/Repository/CommentRepository.php
@@ -1,0 +1,28 @@
+<?php
+declare(strict_types = 1);
+
+namespace FriendsOfTYPO3\BlogExample\Domain\Repository;
+
+/*
+ * This file is part of the TYPO3 CMS project.
+ *
+ * It is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License, either version 2
+ * of the License, or any later version.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ *
+ * The TYPO3 project - inspiring people to share!
+ */
+
+use TYPO3\CMS\Extbase\Persistence\QueryInterface;
+use TYPO3\CMS\Extbase\Persistence\Repository;
+
+/**
+ * A repository for persons
+ */
+class CommentRepository extends Repository
+{
+    protected $defaultOrderings = ['date' => QueryInterface::ORDER_DESCENDING];
+}
