+++ Configuration/Backend/Modules.php
@@ -1,9 +1,9 @@
 <?php
-use FriendsOfTYPO3\BlogExample\Controller\BackendBlogController;
-use FriendsOfTYPO3\BlogExample\Controller\BackendPostController;
-use FriendsOfTYPO3\BlogExample\Controller\CommentController;
+
+use FriendsOfTYPO3\BlogExample\Controller\BackendController;
+
 /**
- * Definitions for modules provided by EXT:examples
+ * Definitions for the backend module provided by EXT:blog_example
  */
 return [
     'blog_example' => [
@@ -15,11 +15,13 @@ return [
         'labels' => 'LLL:EXT:blog_example/Resources/Private/Language/Module/locallang_mod.xlf',
         'extensionName' => 'BlogExample',
         'controllerActions' => [
-            BackendBlogController::class => [
-                'index','deleteAll','populate',
-            ],
-            BackendPostController::class => [
-                'index', 'show',
+            BackendController::class => [
+                'index',
+                'deleteAll',
+                'populate',
+                'showBlog',
+                'showPost',
+                'showAllComments',
             ],
         ],
     ],
