+++ Resources/Private/Templates/Backend/Index.html
@@ -13,7 +13,7 @@
                 <f:for each="{blogs}" as="blog">
                     <li>
                         <h3>
-                            <f:link.action action="index" controller="BackendPost" arguments="{blog: blog}" title="Edit this Blog">
+                            <f:link.action action="showBlog" controller="Backend" arguments="{blog: blog}" title="Edit this Blog">
                                 <core:icon identifier="actions-document-edit" />
                                 {blog.title} (
                                 <f:count subject="{blog.posts}" />
