+++ Resources/Private/Templates/Backend/ShowAllComments.html
@@ -1,0 +1,35 @@
+<html
+    xmlns:f="http://typo3.org/ns/TYPO3/CMS/Fluid/ViewHelpers"
+    xmlns:core="http://typo3.org/ns/TYPO3/CMS/Core/ViewHelpers"
+    data-namespace-typo3-fluid="true">
+
+<f:layout name="Module" />
+
+<f:section name="Content">
+    <h2>List of all comments &quot;{blog.title}&quot;</h2>
+    <div class="tx-blogexample-list-container">
+        <f:if condition="{comments}">
+            <f:then>
+                <ul>
+                    <f:for each="{comments}" as="comment">
+                        <li>
+                            <h3>
+                                <f:link.action action="show" controller="BackendPost" arguments="{post: post, blog: post.blog}">
+                                    <core:icon identifier="actions-document-edit" />
+                                    {post.title}
+                                </f:link.action>
+                            </h3>
+                            <p class="bodytext">
+                                <f:format.crop maxCharacters="100">{post.content}</f:format.crop>
+                            </p>
+                        </li>
+                    </f:for>
+                </ul>
+            </f:then>
+            <f:else>
+                <p>This blog currently doesn't contain any posts.</p>
+            </f:else>
+        </f:if>
+    </div>
+</f:section>
+</html>
